@@ -13,7 +13,7 @@ import java.net.URISyntaxException;
 import java.util.Base64;
 
 public class Main {
-    public static void main(String[] args) throws URISyntaxException, IOException {
+    public static void main(String[] args) throws IOException {
         String apiToken = Base64
                 .getEncoder()
                 .encodeToString(args[0].getBytes());
@@ -23,7 +23,7 @@ public class Main {
         OkHttpClient okHttpClient = new OkHttpClient();
 
         Request requestGetTask = new Request.Builder()
-                .url("https://utrace.atlassian.net/rest/api/2/issue/UTH-4337")
+                .url("https://utrace.atlassian.net/rest/api/3/issue/UTH-4337/changelog")
                 .get()
                 .addHeader("Authorization", authValue)
                 .build();
@@ -35,6 +35,7 @@ public class Main {
         ObjectMapper objectMapper = new ObjectMapper()
                 .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
 
+        assert responseBody != null;
         Issue issue = objectMapper.readValue(responseBody.string(), Issue.class);
 
         System.out.println(responseBody.string());
